GAUD-8854: Make offscreen styles node safe
