+++ components/offscreen/offscreen.js
@@ -1,27 +1,8 @@
 import { css, html, LitElement } from 'lit';
 import { getFlag } from '../../helpers/flags.js';
+import { _offscreenStyleDeclarations, offscreenStyles } from './offscreen-styles.js';
 
-const OffSCREEN_SIZELESS = getFlag('d2l-offscreen-sizeless', true);
-
-/**
- * A private helper declarations that should not be used by general consumers
- */
-export const _offscreenStyleDeclarations = css`
-	direction: var(--d2l-document-direction, ${document.dir === 'rtl' ? css`rtl` : css`ltr`}); /* stylelint-disable-line @stylistic/string-quotes */
-	height: ${OffSCREEN_SIZELESS ? 0 : 1}px;
-	inset-inline-start: -10000px;
-	overflow: hidden;
-	position: absolute !important;
-	white-space: nowrap;
-	width: ${OffSCREEN_SIZELESS ? 0 : 1}px;
-	${document.dir === 'rtl' ? css`right` : css`left`}: -10000px;
-`;
-
-export const offscreenStyles = css`
-	.d2l-offscreen {
-		${_offscreenStyleDeclarations}
-	}
-`;
+export { _offscreenStyleDeclarations, offscreenStyles };
 
 /**
  * A component for positioning content offscreen to only be visible to screen readers.
